# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: PEPA — Platform Engineering & Pipeline Automator API
  description: |
    PEPA — Platform Engineering & Pipeline Automator API for managing services, workflows, scorecards,
    plugins, RBAC, and platform operations.
  version: "0.1.0"
  contact:
    name: PEPA Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8088/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /role-assignments:
    get:
      tags: [RBAC]
      summary: List role assignments
      operationId: listAssignments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  assignments:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleAssignment"
                  total: { type: integer }
    post:
      tags: [RBAC]
      summary: Assign role to user
      operationId: assignRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, role_id]
              properties:
                user_id: { type: string, format: uuid }
                role_id: { type: string, format: uuid }
      responses:
        "201":
          description: Created
  /role-assignments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [RBAC]
      summary: Revoke role assignment
      operationId: revokeAssignment
      responses:
        "200":
          description: Assignment revoked
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleAssignment:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        role_id: { type: string, format: uuid }
        role_name: { type: string }
        is_active: { type: boolean }